+++ diesel/src/pg/serialize/mod.rs
@@ -1,0 +1,3 @@
+mod write_tuple;
+
+pub use self::write_tuple::WriteTuple;